[REFACTOR] [Notifications] 잘못된 패키지 이름 수정
