+++ app/src/main/java/com/example/android_repo_04/data/dto/notification/Subject.kt
@@ -1,4 +1,4 @@
-package com.example.android_repo_04.data.dto.notificaiton
+package com.example.android_repo_04.data.dto.notification
 
 import com.google.gson.annotations.SerializedName
